requirements.txt
alembic==1.2.1
amqp==2.5.1
aniso8601==8.0.0
appdirs==1.4.3
astroid==2.3.2
atomicwrites==1.3.0
attrs==19.3.0
billiard==3.6.1.0
celery==4.3.0
Click==7.0
coverage==4.5.4
dnspython==1.16.0
eventlet==0.25.1
factory-boy==2.12.0
Faker==2.0.3
Flask==1.1.1
Flask-Celery-Helper==1.1.0
Flask-Login==0.4.1
Flask-Migrate==2.5.2
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
greenlet==0.4.15
gunicorn==19.9.0
importlib-metadata==0.23
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.1.1
kombu==4.6.5
lazy-object-proxy==1.4.2
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
more-itertools==7.2.0
packaging==19.2
pluggy==0.13.0
psycopg2-binary==2.8.3
py==1.8.0
pylint==2.4.3
pyparsing==2.4.2
pyrsistent==0.15.4
pytest==5.2.1
pytest-cov==2.8.1
pytest-cover==3.0.0
pytest-coverage==0.0
pytest-html==2.0.0
pytest-metadata==1.8.0
pytest-sugar==0.9.2
python-dateutil==2.8.0
python-editor==1.0.4
python-engineio==3.9.3
python-socketio==4.3.1
pytz==2019.3
six==1.12.0
SQLAlchemy==1.3.10
termcolor==1.1.0
text-unidecode==1.3
typed-ast==1.4.0
vine==1.3.0
wcwidth==0.1.7
Werkzeug==0.16.0
wrapt==1.11.2
WTForms==2.2.1
zipp==0.6.0
